+++ packages/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @vegastack/pages-core
 
+## 0.1.8
+
 ## 0.1.7
 
 ## 0.1.6